Fixes for picky old gcc compiler (4.1.2)
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@8270 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ pde/src/solver/common/AbstractNonlinearElasticityAssembler.hpp
@@ -579,7 +579,7 @@ double AbstractNonlinearElasticityAssembler<DIM>::UpdateSolutionUsingLineSearch(
         index++;
     }
     
-    double best_index;
+    unsigned best_index;
     
     if(index==damping_values.size() && (next_resid_norm < current_resid_norm))
     {
